make Xcode 8 happy.

+++ socksclient/BSocksClient.c
@@ -198,7 +198,7 @@ void connector_handler (BSocksClient* o, int is_error)
     }
     
     // send
-    PacketPassInterface_Sender_Send(o->control.send_if, (uint8_t *)o->buffer, size.value);
+    PacketPassInterface_Sender_Send(o->control.send_if, (uint8_t *)o->buffer, (int) size.value);
     
     // set state
     o->state = STATE_SENDING_HELLO;
@@ -300,7 +300,7 @@ void recv_handler_done (BSocksClient *o, int data_len)
                     ptr += ai->password.password_len;
                     
                     // start sending
-                    PacketPassInterface_Sender_Send(o->control.send_if, (uint8_t *)o->buffer, size.value);
+                    PacketPassInterface_Sender_Send(o->control.send_if, (uint8_t *)o->buffer, (int) size.value);
                     
                     // set state
                     o->state = STATE_SENDING_PASSWORD;
@@ -409,7 +409,7 @@ void send_handler_done (BSocksClient *o)
             }
             
             // receive hello
-            start_receive(o, (uint8_t *)o->buffer, size.value);
+            start_receive(o, (uint8_t *)o->buffer, (int) size.value);
             
             // set state
             o->state = STATE_SENT_HELLO;
@@ -444,7 +444,7 @@ void send_handler_done (BSocksClient *o)
             }
             
             // receive reply header
-            start_receive(o, (uint8_t *)o->buffer, size.value);
+            start_receive(o, (uint8_t *)o->buffer, (int) size.value);
             
             // set state
             o->state = STATE_SENT_PASSWORD;
@@ -499,7 +499,7 @@ void auth_finished (BSocksClient *o)
     memcpy(o->buffer, &header, sizeof(header));
     
     // send request
-    PacketPassInterface_Sender_Send(o->control.send_if, (uint8_t *)o->buffer, size.value);
+    PacketPassInterface_Sender_Send(o->control.send_if, (uint8_t *)o->buffer, (int) size.value);
     
     // set state
     o->state = STATE_SENDING_REQUEST;
